CLN move lista_analysis_dual to its own file
